Move pb conversion modules up one level
