+++ src/core/ingestor/mod.rs
@@ -1,6 +1,6 @@
 pub mod config;
+pub mod conversion;
 pub mod errors;
-pub mod pb;
 pub mod task;
 
 pub use crate::pb::ingestor::PbAprsPacket;
