--- src/core/ingestor/pb/mod.rs
@@ -1,1 +1,0 @@
-pub mod conversion;
